requirements.txt
beautifulsoup4
requests
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
